Move il2p_test.c back into src, with some ifdef guards
DW uses selective linking to manage mocking for tests, cross-platform
logistics, etc. - so multiple functions of the same name across files in
src, with CMake managing the right selection for each binary

This doesn't work so nicely with cgo, which wants to link everything
together in one big blob.

All the main()s were easy to just rename - here I'm checking UNITTEST
and redefining a name (this probably doesn't actually work when UNITTEST
is defined, but I'll fix that later)