Fix decimate settings for atest
I misread upstream in the port of argument parsing - the default was 0
but you could only set 1-8 from the CLI - that was easier to do with the
C argparse setup. So let's allow 0-8 and use the correct default.

This caused a fun issue with check-modem300 where we got subtly
different decode results from upstream, because we were using decimate=1
and they used decimate=0 which became 3 (see src/demod.c)